refactor: 선호 물품 영속성 레이어 인터페이스화 진행
- 기존 Jpa 단일 레포지토리에서 인터페이스, 구현체 Jpa 레포지토리로 분할
- 분할에 따른 의존성 경로 수
