+++ src/main/java/com/somemore/center/repository/preferitem/PreferItemRepository.java
@@ -1,15 +1,13 @@
-package com.somemore.center.repository;
+package com.somemore.center.repository.preferitem;
 
 import com.somemore.center.domain.PreferItem;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.UUID;
 
 @Repository
-public interface PreferItemRepository extends JpaRepository<PreferItem, Long> {
-
+public interface PreferItemRepository {
+    void save(PreferItem preferItem);
     List<PreferItem> findByCenterId(UUID centerId);
-
 }
