+++ src/main/java/com/somemore/center/service/command/UpdateCenterProfileService.java
@@ -2,7 +2,7 @@ package com.somemore.center.service.command;
 
 import com.somemore.center.domain.Center;
 import com.somemore.center.dto.request.CenterProfileUpdateRequestDto;
-import com.somemore.center.repository.CenterRepository;
+import com.somemore.center.repository.center.CenterRepository;
 import com.somemore.center.usecase.command.UpdateCenterProfileUseCase;
 import com.somemore.global.exception.BadRequestException;
 import lombok.RequiredArgsConstructor;
