+++ src/test/java/com/somemore/center/service/query/CenterSignServiceTest.java
@@ -2,7 +2,7 @@ package com.somemore.center.service.query;
 
 import com.somemore.IntegrationTestSupport;
 import com.somemore.center.domain.Center;
-import com.somemore.center.repository.CenterRepository;
+import com.somemore.center.repository.center.CenterRepository;
 import com.somemore.global.exception.BadRequestException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
