+++ src/test/java/com/somemore/interestcenter/service/CancelInterestCenterServiceTest.java
@@ -2,7 +2,7 @@ package com.somemore.interestcenter.service;
 
 import com.somemore.IntegrationTestSupport;
 import com.somemore.center.domain.Center;
-import com.somemore.center.repository.CenterRepository;
+import com.somemore.center.repository.center.CenterRepository;
 import com.somemore.global.exception.BadRequestException;
 import com.somemore.interestcenter.domain.InterestCenter;
 import com.somemore.interestcenter.dto.request.RegisterInterestCenterRequestDto;
