+++ src/test/java/com/somemore/note/service/SendNoteToCenterServiceTest.java
@@ -2,7 +2,7 @@ package com.somemore.note.service;
 
 import com.somemore.IntegrationTestSupport;
 import com.somemore.center.domain.Center;
-import com.somemore.center.repository.CenterRepository;
+import com.somemore.center.repository.center.CenterRepository;
 import com.somemore.global.exception.BadRequestException;
 import com.somemore.note.domain.Note;
 import com.somemore.note.dto.SendNoteToCenterRequestDto;
